Enable sync if multiple files are opened at once
If multiple cfl's are specified on the command line, they are now all
added to the window list to enable syncing between them.
